insta-clone-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
sqlalchemy==2.0.40
python-multipart==0.0.20
supabase==2.15.1
